[BE] refactor: reviews 변수 int type으로 변경
Issue #26

+++ BE/airbnb5/src/main/java/com/codesquad/airbnb5/dto/RoomDetailDto.java
@@ -35,7 +35,7 @@ public class RoomDetailDto {
     @JsonProperty("scores")
     private final float scores;
     @JsonProperty("reviews")
-    private final float reviews;
+    private final int reviews;
     @JsonProperty("amenities")
     private final String amenities;
     @JsonProperty("cleaning_fee")
